Juntae error (#13)
* feature: make change password button and etc.

* 카톡에 적은 수많은 오류를 해결
