+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.13.0",
+    "@mui/material": "^5.13.2",
     "@reduxjs/toolkit": "^1.9.5",
     "@types/jest": "^29.5.1",
     "@types/node": "20.1.2",
@@ -19,10 +19,10 @@
     "@types/react-dom": "18.2.4",
     "autoprefixer": "10.4.14",
     "axios": "^1.4.0",
-    "next": "^13.4.1",
+    "next": "^13.4.3",
     "postcss": "8.4.23",
     "react": "^18.2.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-redux": "^8.0.5",
     "redux-persist": "^6.0.0",
